+++ src/utils.ts
@@ -1,15 +1,36 @@
 /**
+	* Multiplier to convert degrees into radians.
+	*/
+export const deg2rad = Math.PI / 180;
+
+/**
+ * Multiplier to convert radians into degrees.
+ */
+export const rad2deg = 180 / Math.PI;
+
+/**
+ * Round the number up to the desired precision.
+ */
+export function ceil(number: number, digits = 0, base = Math.pow(10, digits)): number {
+	return Math.ceil(number * base) / base + 0;
+}
+
+/**
+ * Converts the angle in radians to degrees if specified.
+ */
+export const convertAngle = (value: number, degrees = false) => {
+	return degrees
+		? value * rad2deg
+		: value;
+};
+
+/**
  * Clamps a value between an upper and lower bound.
  * NaN is clamped to the lower bound
  */
 export const clamp = (number: number, min = 0, max = 1): number => {
 	return Math.min(Math.max(number, min), max);
 };
-
-/**
-	* Multiplier to convert degrees into radians.
-	*/
-export const deg2rad = Math.PI / 180;
 
 export const isObject = <T = Record<string, unknown>>(input: unknown): input is T => {
 	return typeof input === "object"
@@ -18,9 +39,11 @@ export const isObject = <T = Record<string, unknown>>(input: unknown): input is 
 };
 
 /**
- * Multiplier to convert radians into degrees.
+ * Round the number up to the desired precision.
  */
-export const rad2deg = 180 / Math.PI;
+export function floor(number: number, digits = 0, base = Math.pow(10, digits)): number {
+	return Math.floor(number * base) / base + 0;
+}
 
 /**
  * Round the number up to the desired precision.
@@ -34,13 +57,4 @@ export function round(number: number, digits = 0, base = Math.pow(10, digits)): 
  */
 export const validateNumbers = (...items: unknown[]) => {
 	return items.every(item => typeof item === "number" && !Number.isNaN(item));
-};
-
-/**
- * Converts the angle in radians to degrees if specified.
- */
-export const convertAngle = (value: number, degrees = false) => {
-	return degrees
-		? value * rad2deg
-		: value;
 };